Fixed animation not appearing when using new blur approach
Animating change of an image when an `UIVisualEffectView` is placed above results in no animation. The solution is to take a snapshot, append it to the superview, hide the real view, change it without animation and then fade the real view in with animation.

This post was created in the process:
https://stackoverflow.com/q/59412432/5827933