Use MediaRemote method for the Music app
If only using the MediaRemote method when no artworkIdentifier is present, it results in the next track's artwork sometimes. If only using it for the iOS 13 version (as the iOS 12 version always sends the artworkIdentifier), the iOS 12 version will for some reason result in no animation for some users. Why this happens is beyond me.