Fixed timing inconsistencies between Canvas and Artwork
Inconsistencies would previously happen for example when the Canvas video had to be downloaded. The artwork would be shown briefly only for it to be replaced by the Canvas. Now SpringArtwork instead uses specific code in Spotify to fetch and send the artwork only if no canvas is available (or if Canvas is disabled in SpringArtwork itself).