Revert of DRM move to the manager
It doesn't result in anything more secure from hacks since the init calls need to be executed which are inline methods.

Moreover, having the DRM in the manager implementation results in some hooks not being initialized (in time?) for one user running odysseyra1n, making the lockscreen pulldown not being themed.